extracted out the macros
